Try to detect crashed specs better
Beforehand, we were checking if the spec process exited normally. If not
we assumed that it would have timeouted.

However, some specs exit with an SIGABRT, which means they did not exit
normally. This led to them not being marked as crashed.

With this patch, instead of relying on the status, we use the current
spec details to determine whether the process timed out.