Make Module#define_method more spec compliant
Natalie still cannot run all #define_method specs however we might have
come a little closer :^)

Currently our block handling is pretty limited to blocks being blocks
which they might not be (when passed to a define_method). This is why a
lot of the changes in this commit seem hackish or like a workaround.

It might be possible to separate the actual code of the block and the
special block handling (like raising jump errors) and only use the
actual "code" of the block for defining a custom method when calling
Module#define_method. However, this would require a major refactor of
our current block handling...